generic: remove support for final target via KDE4_ENABLE_FINAL
it is bogus with the CMake automoc, requires additional guards in
the source code and some components do not support it at all. Also,
Link Time Optimizations (LTO) seem to be doing exactly what
KDE4_ENABLE_FINAL was supposed to do (basicly) so no point in
trying to support something that modern compilers can already do
for us.