kio: do not show some metadata fields by default
namely "Part of" and an extra "Type". The "Part of" filed is not
very obvious and in most cases just a null (empty string), the
extra "Type" refers to some sorta Semantic spec thing is not more
than useless as two other MIME information fileds are show - one
with a human friendly and one with programmers friendly string in
it and are more than enough in most if not all cases.