kio: fix crash due to exposing inconsistent views of internal data
From: Simeon Bird <bladud@gmail.com>
Date: Thu, 18 Dec 2014 03:54:32 +0000
Subject: Fix crash in KIO due to exposing inconsistent views of internal data.
X-Git-Tag: v4.14.8
X-Git-Url: http://quickgit.kde.org/?p=kdelibs.git&a=commitdiff&h=4288d7e7c81839d6bdb25b1197c6e614f0259036
---
Fix crash in KIO due to exposing inconsistent views of internal data.

This can be triggered by renaming a directory while one of the files in
it is open in gwenview.

It occurs because when KCoreDirListerCache::emitRedirections is called,
itemsInUse contains the old url. However, KDirLister::Private::redirect
changes lstDirs to the new url. Thus at this point lstDirs contains an
item not in itemsInUse, which causes an assertion if forgetDirs is
called.

In gwenview, the redirection signal is connected to openURL. This calls
forgetDirs, and causes the assertion.

The solution: update itemsInUse *before* emitting redirections.

This fixes the crash, but gwenview opens the first file in
the new directory instead of the file open before renaming.
This is probably an unrelated gwenview bug.

REVIEW: 117345
---