kdecore: limit standard directories cache to 10000 entries
this also fixes a strange bug that I originally tought that it's an RPATH
issues but it turns out that this is some kind of mutex-lock issue. to be
more interesting this bug does not occur when the NDEBUG conditionals are
removed or if they are unconditionalized, the first observation leads me to
think that this is compiler bug and the second that it's mutex-lock issue
that is resolved because some code in KDebug and/or QDebug (i've tried with
it too) locks/unlockes/detaches/does some voodoo that overcomes the bug.

in any case it's hard for me to pin-point the problem so if anyone can look
into it that would be great, for now this change will do


Signed-off-by: Ivailo Monev <xakepa10@gmail.com>