khtml: use QImageReader explicitly
this allows to use char* instead of uchar* to avoid some casts,
also incremental requests are no more and the data is handled in
one go to avoid a lot of small memory allocations and extra array
for each image. it does perform slitely worse but with this a lot
more image formats can be handled with higher quality (slitely)
and even more correct, some images previously did not show
at all because of the broken image handlers.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>