khtml: store page caches in memory
they are small enough (usually ~200KB) to not cause major increase
of system requirements as QCache is used and it defaults to not
more than 10 entries. worst case scenario they can be compressed
via qCompress() but that will make things a bit more complicated.

the change prevents exposure of the page caches so that
overwriting them for malicious purposes is less likely to happen.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>