khtml: simplify line break handling
for the Thai language it get's it right most of the time, the
point is to avoid dlopen-ing the library and resolving the symbols
at runtime avoiding overheat and dependency on the library by
making use of what Qt4/Katie offers. the actual compile-time check
for libthai was not in place so it was always probed for at
runtime which can cause all kinds of problems and leaves open an
attack vector.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>