kutils: improve the position slider and hidden controls behaviour
now when the slider is being dragged the player does not continuesly seek,
it does it only when the slider is released to avoid bad noise during the
quick seeking.

also, the controls will not be hidden now while any of the controls is
being pressed by the user, it's still not perfect because once the user
releases it the controls will be hidden and then (most likely) shown again
(unless no mouse was used for the interactive, e.g. tab for selection and
then space for interaction) but that will require even more connections
to the controls so I'm leaving it as is for now.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>