kutils: add kaudioplayer KDED module
the module has some use cases, like avoiding linkage to the
kmediaplayer library which links to MPV, which links to FFmpeg.
and since the module is loaded on demand there should be no change
in terms of resources usage untils it is used, in which case it
will use (roughly) the resources of single MPV instance until
kded4 quits.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>