cmake: MacroLogFeature rewrite
logging found and otherwise packages will no longer be done thus
macro_display_feature_log() is removed. the reason for not doing is because
it simply does not cover all package lookups, only those logged via
macro_log_feature(). if I was to fix that, I would probably resort to
CMakeCache.txt parsing trickery since it includes indirect package lookups
aswell as header, struct, member, etc. checks however that would be
fragile and require common pattern in the checks

macro_log_feature() incorrectly quoted DESCRIPTION and PURPOSE in the
CMake output but that is fixed now. it will be replaced with actual
set_package_properties() macro calls in the future.

in any case, it is addition disk I/O so the configuration/build time will
be reduced with this change.

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>