cmake: rewrite Qalculate module
neither the Plasma runner nor the applet actually used the QALCULATE_CFLAGS
variable meaning that if Qalculate was installed in non-standard prefix
the build was likely to fail, this fixes that problem by replacing it with
QALCULATE_INCLUDE_DIR which will be used in the build system for the applet
and runner.

also fixes a CMake deprecation warning about using obsolete UsePkgConfig
module.

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>