cmake: rewrite PCIUTILS module
it was causing overlinkg due to incorrect dependency additions added by the
module because resolv library is used privately. in fact, it uses zlib and
udev too but only privately and none of their include directories are
needed to use libpci.

since the rewrite makes use of pkg-config as primary lookup method any
future changes to that behaviour will be handled correctly without
overlinking.

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>