generic: add new signal to solid device manager
its use case is shown in KFilePlacesModel, basically when a CD-ROM is
ejected it will be removed from the items in the model and when disc is
inserted item for it will be added to the model. Dolphin and Plasma require
changes of their own for this behaviour to be consistent across
applications

naming the method after OpticalDisc::availableContent(), have in mind it
should work when a disc is erased and written to for example or when a
device is formatted and no longer has filesystem which has not been tested
but in theory should work. may need additional checks for encrypted device
but that can be done in the future

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>