kutils: change the slider position before load() in KMediaWidget::open()
fixes rare runtime warning about seek command failure. the proper thing
to do is close, wait for close to happen, set slider position, load and
continue the event cycle but that's too complicated when buffering is not
expected to happen on seek (i.e. when a file from the hard-disk is
currently loaded)

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>