kdeui: quit KApplication when D-Bus daemon disconnects
so that applications (services like kglobalaccel, knotify, etc.) do not
keep running after the daemon quits (e.g. when program is started via
`kdesudo`, daemon quits but does not terminate services). there is a catch
tho - the order in which services quit may cause some trouble for
applications that do not expect services to suddenly stop

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>