kio: magick bits to watch the URL of desktop files from KDirLister
this essentially will watch the .desktop file URL which is, say, "trash:/"
and emit changes as if the .desktop file itself was modified (to update the
icon for example). that's what the KIO desktop slave was doing essentially
but now it is implemented into KDirLister itself so anything using that
class will get it!

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>